Refactor WikiCitaitons class into WikibaseCrud and 3 subclasses WikibaseCrudRead, WikibaseCrudCreate, WikibaseCrudDelte. __init__.py: WikibaseCrud: Renamed and moved class __init__.py: WcdImportBot: delete_one_page(), rinse_all_items_and_cache(): Use WikibaseCrudDelete. lookup_md5hash(): Use WikibaseCrudRead create.py, delete.py, read.py: New files enums.py: Move file setup_all_properties_and_items_on_new_wikibase.py: __setup_archive_items__(): Update class name test_*: Update to use the new classes wikipedia_page.py, wikipedia_page_reference.py: Update class name
